+++ hydroflow/src/lib.rs
@@ -3,7 +3,6 @@
 #![cfg_attr(feature = "variadic_generics", feature(generic_associated_types))]
 #![allow(clippy::let_and_return)]
 
-pub mod builder;
 pub mod compiled;
 pub mod lang;
 pub mod scheduled;
